GEN-1124 Add params to bypass some mlnx_ofed tasks
Some appliances like the NVIDIA DGX run their own software stack. When
an appliance already manages the installation of the Mellanox OFED
drivers, there is no need to configure additional repositories or to
install those drivers.

Add two inventory parameters `mlnx_ofed_configure_repos` and
`mlnx_ofed_install_kernel_modules` which allow to bypass configuration
of the packages repositories and installation of the kernel modules.

Default is still to configure the repos and install the kernel modules
during the play.