feat: Add support of SR-IOV configuration in HCA
The role can configure the parameters SRIOV_EN and NUM_OF_VFS in the
Mellanox HCAs. It relies on the command `mstconfig` and the PCI address
of the devices to define the parameters. These must be defined in the
inventory with the optional variable `infiniband_hca_devices`. An
example is provided in the defaults.

If the variable is not defined, the role will skip the configuration of
the HCA.

A reboot is mandatory after a configuration change on the HCA. Add a new
handler to reboot the host. Since SR-IOV is mostly used for hypervisors,
make sure to serialize the reboots with `throttle: 1`. This will allow
to reboot the hosts one after the other, with limited impact on services
hosted in virtualization clusters.