feat: configure SR-IOV on the hosts
It is usually possible to create up to NUM_OF_VFS in the kernel by
writing the requested number of VFs in the `sriov_numvfs` Linux kernel
file.

In order to ease the configuration, this role will set `sriov_numvfs =
NUM_OF_VFS`, so the maximum of VFs will be created. If one needs less
VFs, they will have to lower NUM_OF_VFS as well. The is configurable
with `infiniband_hca_devices[*].num_of_vfs` in the inventory.

Install the sysfsutils package and write the sysfs configuration file to
make the configuration persistent on reboot.

Define the VF GUID from a prefix + device ID + VF ID.