fix: Do not force reboot to load new HCA configs
In some cases, one may want to apply a new configuration to an HCA
without rebooting the host. Add a new variable `infiniband_allow_reboot`
that permits to disable the reboot that should occur after an HCA
configuration change.

Also allow to increase the default throttle of 1 with
`infiniband_throttle_reboot`.