fix: Allow to unload irdma with openibd
It might happen that the in-kernel module ib_uverbs is in use by the
in-kernel module irdma on servers with Ethernet interfaces that support
RDMA. In order to reload the MLNX_OFED modules after the installation,
openibd must unload the module irdma.

Add a mechanism which allows to apply known patches on demand with the
role.